requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1.6
django-autoslug==1.9.9
django-crispy-forms==2.1
djangorestframework==3.14.0
psycopg>=3.1.0
python-dotenv==1.2.1
sqlparse==0.5.0
